Revert "refactor(select): MenuItem -> Option"
This reverts commit 388ef58eb972111dd821efd2a650f94bb6e13f23.
