+++ packages/select/src/Component.stories.mdx
@@ -3,7 +3,7 @@ import { Meta, Story, Props, Title } from '@storybook/addon-docs/blocks';
 import { Select } from './Component';
 import { Field } from './components/Field';
 import { Menu } from './components/Menu';
-import { Option } from './components/Option';
+import { MenuItem } from './components/MenuItem';
 import { name, version } from '../package.json';
 
 import Icon from '@alfalab/icons-classic/BankAlfaMBlackIcon';
@@ -169,7 +169,7 @@ import { Select } from '@alfalab/core-components-select';
 
 <Props of={Menu} />
 
-## Option Props
+## MenuItem Props
 
-<Props of={Option} />
+<Props of={MenuItem} />
