+++ packages/select/src/components/MenuItem/index.module.css
@@ -1,6 +1,6 @@
 @import '../../../../themes/src/default.css';
 
-.option {
+.item {
     @mixin system_14-18_regular;
 
     display: block;
